+++ Reports/CHANGES.md
@@ -1,6 +1,12 @@
 # CHANGES
 ## Reports module for RosarioSIS
 
+Changes in 1.3
+--------------
+- Compatibility with RosarioSIS 2.9 in SavedReports.php
+- Use Bottom.php|bottom_buttons hook for Save Report button in functions.php
+
+
 Changes in 1.2
 --------------
 - Adjust Search screens position and reset Equation on Save in Calculations.php & functions.js